+++ ding-impl/ding-presentation-impl/src/main/java/org/cytoscape/ding/internal/charts/bar/BarChart.java
@@ -87,9 +87,9 @@ public class BarChart extends AbstractChartCustomGraphics<BarLayer> {
 		final boolean showLabels = get(SHOW_ITEM_LABELS, Boolean.class, false);
 		final boolean showDomainAxis = get(SHOW_DOMAIN_AXIS, Boolean.class, false);
 		final boolean showRangeAxis = get(SHOW_RANGE_AXIS, Boolean.class, false);
-		final double axisWidth = get(AXIS_WIDTH, Double.class, 0.25);
+		final float axisWidth = get(AXIS_WIDTH, Float.class, 0.25f);
 		final Color axisColor = get(AXIS_COLOR, Color.class, Color.DARK_GRAY);
-		final double borderWidth = get(BORDER_WIDTH, Double.class, 0.25);
+		final float borderWidth = get(BORDER_WIDTH, Float.class, 0.25f);
 		final Color borderColor = get(BORDER_COLOR, Color.class, Color.DARK_GRAY);
 		
 		double separation = get(SEPARATION, Double.class, 0.0);
