+++ ding-impl/ding-presentation-impl/src/main/java/org/cytoscape/ding/internal/charts/line/LineChart.java
@@ -76,7 +76,7 @@ public class LineChart extends AbstractChartCustomGraphics<LineLayer> {
 		final boolean showItemLabels = get(SHOW_ITEM_LABELS, Boolean.class, false);
 		final boolean showDomainAxis = get(SHOW_DOMAIN_AXIS, Boolean.class, false);
 		final boolean showRangeAxis = get(SHOW_RANGE_AXIS, Boolean.class, false);
-		final double axisWidth = get(AXIS_WIDTH, Double.class, 0.25);
+		final float axisWidth = get(AXIS_WIDTH, Float.class, 0.25f);
 		final Color axisColor = get(AXIS_COLOR, Color.class, Color.DARK_GRAY);
 		final float lineWidth = get(LINE_WIDTH, Float.class, 1.0f);
