+++ ding-impl/ding-presentation-impl/src/main/java/org/cytoscape/ding/internal/charts/pie/PieChart.java
@@ -65,7 +65,7 @@ public class PieChart extends AbstractChartCustomGraphics<PieLayer> {
 		final Rectangle2D bounds = new Rectangle2D.Double(-size / 2, -size / 2, size, size);
 		final boolean showLabels = get(SHOW_ITEM_LABELS, Boolean.class, false);
 		
-		final double borderWidth = get(BORDER_WIDTH, Double.class, 0.25);
+		final float borderWidth = get(BORDER_WIDTH, Float.class, 0.25f);
 		final Color borderColor = get(BORDER_COLOR, Color.class, Color.DARK_GRAY);
 		
 		final PieLayer layer = new PieLayer(data, labels, showLabels, colors, borderWidth, borderColor, startAngle,
