+++ ding-impl/ding-presentation-impl/src/main/java/org/cytoscape/ding/internal/charts/ring/RingChart.java
@@ -67,7 +67,7 @@ public class RingChart extends AbstractChartCustomGraphics<RingLayer> {
 		final Rectangle2D bounds = new Rectangle2D.Double(-size / 2, -size / 2, size, size);
 		final boolean showLabels = get(SHOW_ITEM_LABELS, Boolean.class, false);
 		
-		final double borderWidth = get(BORDER_WIDTH, Double.class, 0.25);
+		final float borderWidth = get(BORDER_WIDTH, Float.class, 0.25f);
 		final Color borderColor = get(BORDER_COLOR, Color.class, Color.DARK_GRAY);
 		
 		final RingLayer layer = new RingLayer(data, labels, showLabels, colors, borderWidth, borderColor, startAngle,
