+++ ding-impl/ding-presentation-impl/src/main/java/org/cytoscape/ding/internal/charts/stripe/StripeChart.java
@@ -74,7 +74,7 @@ public class StripeChart extends AbstractChartCustomGraphics<StripeLayer> {
 		final double size = 32;
 		final Rectangle2D bounds = new Rectangle2D.Double(-size / 2, -size / 2, size, size);
 		
-		final double borderWidth = get(BORDER_WIDTH, Double.class, 0.25);
+		final float borderWidth = get(BORDER_WIDTH, Float.class, 0.25f);
 		final Color borderColor = get(BORDER_COLOR, Color.class, Color.DARK_GRAY);
 		
 		final StripeLayer layer = new StripeLayer(data, labels, false, colors, borderWidth, borderColor,
